[RFXCOM] Receive-only devices
Add thing config options to allow a device to be set receive-only. Many
sensors and wireless switches fall into this category and sending to
them is simply a waste of time. This is especially a problem when
persistence is trying to restore state as the wasted work means it may
take many seconds or even minutes before sanity is restored.

Closes #2287

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk> (github: mjagdis)