[lightify] Improved transitions and state<->item mapping
Suppress item updates during transitions. The state given by LIST_PAIRED
only changes when the gateway interacts with a device or when devices
are periodically polled (every 30s for current firmware). The state
known by the gateway is NOT continuously updated during a transition. If
we do not ignore it we will update items with stale device state from
some unknown milliseconds after the transition was initiated - and if
items are set to autoupdate (the default) then they already know better!

Improve the way we map state changes to item updates so that items are
kept in better sync, taking account of such niceties as openHAB
converting ON and OFF on dimmers to 0 and 100.

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk> (github: mjagdis)